chore: rename api-error to unleash-error
