+++ src/lib/error/feature-has-tag-error.ts
@@ -1,4 +1,4 @@
-import { UnleashError } from './api-error';
+import { UnleashError } from './unleash-error';
 
 class FeatureHasTagError extends UnleashError {}
 export default FeatureHasTagError;
