+++ src/lib/error/invalid-operation-error.ts
@@ -1,4 +1,4 @@
-import { UnleashError } from './api-error';
+import { UnleashError } from './unleash-error';
 
 class InvalidOperationError extends UnleashError {}
 export default InvalidOperationError;
