+++ src/lib/error/invalid-token-error.ts
@@ -1,4 +1,4 @@
-import { UnleashError } from './api-error';
+import { UnleashError } from './unleash-error';
 
 class InvalidTokenError extends UnleashError {
     constructor() {
