+++ src/lib/error/minimum-one-environment-error.ts
@@ -1,4 +1,4 @@
-import { UnleashError } from './api-error';
+import { UnleashError } from './unleash-error';
 
 class MinimumOneEnvironmentError extends UnleashError {}
