+++ src/lib/error/name-exists-error.ts
@@ -1,4 +1,4 @@
-import { UnleashError } from './api-error';
+import { UnleashError } from './unleash-error';
 
 class NameExistsError extends UnleashError {}
 export default NameExistsError;
