+++ src/lib/error/notfound-error.ts
@@ -1,4 +1,4 @@
-import { UnleashError } from './api-error';
+import { UnleashError } from './unleash-error';
 
 class NotFoundError extends UnleashError {
     constructor(message: string = 'The requested resource could not be found') {
