+++ src/lib/error/password-mismatch.ts
@@ -1,4 +1,4 @@
-import { UnleashError } from './api-error';
+import { UnleashError } from './unleash-error';
 
 class PasswordMismatch extends UnleashError {
     constructor(message: string = 'Wrong password, try again.') {
